Create ConditionAwaiter via factory
