+++ awaitility/src/main/java/org/awaitility/core/CallableCondition.java
@@ -39,10 +39,9 @@ class CallableCondition implements Condition<Void> {
     CallableCondition(final Callable<Boolean> matcher, ConditionSettings settings) {
         conditionEvaluationHandler = new ConditionEvaluationHandler<Object>(null, settings);
         ConditionEvaluationWrapper conditionEvaluationWrapper = new ConditionEvaluationWrapper(matcher, settings, conditionEvaluationHandler);
-        conditionAwaiter = new ConditionAwaiter(conditionEvaluationWrapper, settings) {
-            @SuppressWarnings("rawtypes")
+        conditionAwaiter = ConditionAwaiterFactory.getInstance().newConditionAwaiter(conditionEvaluationWrapper, settings, new TimeoutMessageSupplier() {
             @Override
-            protected String getTimeoutMessage() {
+            public String getTimeoutMessage() {
                 if (timeout_message != null) {
                     return timeout_message;
                 }
@@ -67,7 +66,7 @@ class CallableCondition implements Condition<Void> {
                 }
                 return timeoutMessage;
             }
-        };
+        });
     }
 
     /**
