+++ awaitility/src/main/java/org/awaitility/core/ConditionAwaiterDefaultFactory.java
@@ -1,0 +1,12 @@
+package org.awaitility.core;
+
+public class ConditionAwaiterDefaultFactory extends ConditionAwaiterFactory {
+    public ConditionAwaiterDefaultFactory() {
+    }
+
+    public ConditionAwaiter newConditionAwaiter(ConditionEvaluator conditionEvaluator,
+                                                ConditionSettings conditionSettings,
+                                                TimeoutMessageSupplier timeoutMessageSupplier) {
+        return new ConditionAwaiterImpl(conditionEvaluator, conditionSettings, timeoutMessageSupplier);
+    }
+}
