+++ awaitility/src/main/java/org/awaitility/core/TimeoutMessageSupplier.java
@@ -1,0 +1,5 @@
+package org.awaitility.core;
+
+public interface TimeoutMessageSupplier {
+    String getTimeoutMessage();
+}
